backend/pyproject.toml
# pyproject.toml
[project]
name = "llm-webrtc"
version = "0.1.0"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aioice==0.10.1",
    "aiortc==1.13.0",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "av==14.4.0",
    "bidict==0.23.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cryptography==46.0.1",
    "cycler==0.12.1",
    "dnspython==2.8.0",
    "fastapi==0.117.1",
    "filelock==3.20.0",
    "fonttools==4.60.1",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "google-crc32c==1.7.1",
    "h11==0.16.0",
    "httptools==0.6.4",
    "idna==3.10",
    "ifaddr==0.2.0",
    "jinja2==3.1.6",
    "kiwisolver==1.4.9",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "networkx==3.5",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "opencv-python-headless==4.12.0.88",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.0.0",
    "polars==1.35.2",
    "polars-runtime-32==1.35.2",
    "propcache==0.4.1",
    "psutil==7.1.3",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pyee==13.0.0",
    "pylibsrtp==0.12.0",
    "pyopenssl==25.3.0",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-engineio==4.12.2",
    "python-multipart==0.0.20",
    "python-socketio==5.13.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.5",
    "scipy==1.16.3",
    "seaborn==0.13.2",
    "setuptools==80.9.0",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.48.0",
    "sympy==1.14.0",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "torchvision==0.23.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "ultralytics==8.3.229",
    "ultralytics-thop==2.0.18",
    "urllib3==2.5.0",
    "uvicorn==0.37.0",
    "watchfiles==1.1.0",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "wsproto==1.2.0",
    "yarl==1.22.0",
]
